rowexec: move defaultSpanGenerator to joinreader_span_generator.go
Move the defaultSpanGenerator to joinreader_span_generator.go to consolidate
all implementations of joinReaderSpanGenerator in the same file.

Release note: None
